move addNavigation to a module
