+++ index.js
@@ -1,4 +1,5 @@
 import BookCollection from './modules/bookCollection.js';
+import addNavigation from './modules/addNavigation.js';
 
 const bookListWrapper = document.getElementById('bookList');
 const addBookForm = document.getElementById('addBookForm');
@@ -36,20 +37,4 @@ const navigations = [
   },
 ];
 
-navigations.forEach((navigation) => {
-  const link = document.getElementById(navigation.linkId);
-
-  link.addEventListener('click', (event) => {
-    event.preventDefault();
-
-    // hide all sections
-    const allSections = document.querySelectorAll('main > section');
-    allSections.forEach((section) => {
-      section.classList.add('display-none');
-    });
-
-    // show the relevant section
-    const target = document.getElementById(navigation.targetId);
-    target.classList.remove('display-none');
-  });
-});
+addNavigation(navigations);
